requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
weaviate-client==4.12.0
neo4j==5.28.0
prometheus-fastapi-instrumentator
slowapi
pytest
pytest-cov
pytest-asyncio
ruff==0.4.0
black
pydantic==2.10.6
openai>=1.25.0
httpx
apscheduler
python-jose
cachetools
pillow
pymupdf
pytesseract
openpyxl>=3.1.5
reflex==0.7.12 